+++ noxfile.py
@@ -2,7 +2,7 @@
 import nox_poetry
 
 
-@nox_poetry.session()  # for local testing, ex: python=["3.12", "3.13"]
+@nox_poetry.session()
 def all_tests(session):
     """Run the test suite using pytest."""
     session.install("pytest", ".")  # for coverage report include: "pytest-cov"
@@ -14,11 +14,71 @@ def all_tests(session):
     )
 
 
-@nox_poetry.session()  # python=["3.12", "3.13"]
-def single_test(session):
-    """Run a single test using pytest."""
+@nox_poetry.session()
+def evaluations(session):
+    """Run tests related to evaluations."""
     session.install("pytest", ".")
     session.run(
         "pytest",
-        "tests/test_clone_from_s3.py"
+        "tests/evaluations"
+    )
+
+
+@nox_poetry.session()
+def fetch(session):
+    """Run tests related to fetching."""
+    session.install("pytest", ".")
+    session.run(
+        "pytest",
+        "tests/fetch"
+    )
+
+
+@nox_poetry.session()
+def generate(session):
+    """Run tests related to generating timeseries."""
+    session.install("pytest", ".")
+    session.run(
+        "pytest",
+        "tests/generate"
+    )
+
+
+@nox_poetry.session()
+def load(session):
+    """Run tests related to loading."""
+    session.install("pytest", ".")
+    session.run(
+        "pytest",
+        "tests/load"
+    )
+
+
+@nox_poetry.session()
+def query(session):
+    """Run tests related to querying."""
+    session.install("pytest", ".")
+    session.run(
+        "pytest",
+        "tests/query"
+    )
+
+
+@nox_poetry.session()
+def s3_utils(session):
+    """Run tests related to S3 utilities."""
+    session.install("pytest", ".")
+    session.run(
+        "pytest",
+        "tests/s3_utils"
+    )
+
+
+@nox_poetry.session()
+def visualization(session):
+    """Run tests related to visualization."""
+    session.install("pytest", ".")
+    session.run(
+        "pytest",
+        "tests/visualization"
     )
