+++ tests/fetch/test_fetch_and_load.py
