+++ tests/fetch/test_generate_weights.py
