+++ tests/fetch/test_nwm_fetch_and_format.py
