+++ tests/load/test_import_location_attributes.py
