+++ tests/load/test_import_locations.py
