+++ tests/load/test_merge_mapping_fields.py
