+++ tests/visualization/test_dataframe_accessor.py
